requirements.txt
black==24.4.2
Django==5.0.4
djangorestframework==3.15.1
flake8==7.0.0
sqlparse==0.5.0
psycopg2
graphene-django
